+++ inst/NEWS.md
@@ -1,5 +1,31 @@
-# CHANGES IN VERSION 2.12.05
+# CHANGES IN VERSION 2.17.0
 (GitHub master branch)
+
+## BUG FIXES
+- Bug fix in `subsetMaf` while handling only CNV events. [908](https://github.com/PoisonAlien/maftools/issues/908)
+- Error handling when no deep/shallow CNV events found. [899](https://github.com/PoisonAlien/maftools/issues/899)
+- Bug fix in `oncoplot` for duplicated values in gene list. [889](https://github.com/PoisonAlien/maftools/issues/889)
+
+## ENHANCEMENTS
+- Added `showOnlyPathway` argument to `oncoplot`
+- Added `pathdb` argument to `PlotOncogenicPathways`. Issue: [923](https://github.com/PoisonAlien/maftools/issues/923)
+- Emit warnings when fishers test can not be performed during `somaticInteractions`. Issue: [921](https://github.com/PoisonAlien/maftools/issues/921)
+- Added `leftMar` and `topMar` arguments to `somaticInteractions`. Issue: [913](https://github.com/PoisonAlien/maftools/issues/913)
+- Added `toptBarLims` argument to oncoplot. Issue: [910](https://github.com/PoisonAlien/maftools/issues/910)
+- Added `data` argument to `lollipopPlot` function. Issue: [894](https://github.com/PoisonAlien/maftools/issues/894)
+- Added `sortByM1` and `sortByM2` argument to `coOncoplot`. Issue: [888](https://github.com/PoisonAlien/maftools/issues/888)
+- Added arguments `leftBarVline`, `leftBarVlineCol`, `rightBarVline`, `rightBarVlineCol` `topBarHline` `topBarHlineCol` to `oncoplot`. Issue: [874](https://github.com/PoisonAlien/maftools/issues/874)
+- Added `revPal` argument to `somaticInteractions`. Issue: [859](https://github.com/PoisonAlien/maftools/issues/859)
+
+# CHANGES IN VERSION 2.16.0
+(Bioconductor release version)
+
+# CHANGES IN VERSION 2.14.0
+(Bioconductor release version)
+
+- No changes
+
+# CHANGES IN VERSION 2.12.05
 
 ## BUG FIXES
 - genes parameter in `plotClusters` can't accept list of genes. Issue: [828](https://github.com/PoisonAlien/maftools/issues/828)